Fix NULL pointer constant issues in kernel/arch/x86
illumos bug# 6609, 6611

+++ kernel/arch/x86/platform/i86pc/os/cmi_hw.c
@@ -709,7 +709,8 @@ ntv_getcr4(cmi_hdl_impl_t *hdl)
 	cpu_t *cp = HDLPRIV(hdl);
 	ulong_t val;
 
-	(void) call_func_ntv(cp->cpu_id, ntv_getcr4_xc, (xc_arg_t)&val, NULL);
+	(void) call_func_ntv(cp->cpu_id, ntv_getcr4_xc, (xc_arg_t)&val,
+	    (uintptr_t)NULL);
 
 	return (val);
 }
@@ -732,7 +733,8 @@ ntv_setcr4(cmi_hdl_impl_t *hdl, ulong_t val)
 {
 	cpu_t *cp = HDLPRIV(hdl);
 
-	(void) call_func_ntv(cp->cpu_id, ntv_setcr4_xc, (xc_arg_t)val, NULL);
+	(void) call_func_ntv(cp->cpu_id, ntv_setcr4_xc, (xc_arg_t)val,
+	    (uintptr_t)NULL);
 }
 
 volatile uint32_t cmi_trapped_rdmsr;
@@ -839,7 +841,8 @@ ntv_int(cmi_hdl_impl_t *hdl, int int_no)
 {
 	cpu_t *cp = HDLPRIV(hdl);
 
-	(void) call_func_ntv(cp->cpu_id, ntv_int_xc, (xc_arg_t)int_no, NULL);
+	(void) call_func_ntv(cp->cpu_id, ntv_int_xc, (xc_arg_t)int_no,
+	    (uintptr_t)NULL);
 }
 
 static int
@@ -1595,6 +1598,7 @@ cmi_hdl_class(cmi_hdl_t ophdl)
 		    cmio_##what(IMPLHDL(ophdl)));		\
 	}
 
+/* BEGIN CSTYLED */
 CMI_HDL_OPFUNC(vendor, uint_t)
 CMI_HDL_OPFUNC(vendorstr, const char *)
 CMI_HDL_OPFUNC(family, uint_t)
@@ -1614,6 +1618,7 @@ CMI_HDL_OPFUNC(logical_id, id_t)
 CMI_HDL_OPFUNC(smbiosid, uint16_t)
 CMI_HDL_OPFUNC(smb_chipid, uint_t)
 CMI_HDL_OPFUNC(smb_bboard, nvlist_t *)
+/* END CSTYLED */
 
 boolean_t
 cmi_hdl_is_cmt(cmi_hdl_t ophdl)
@@ -1702,8 +1707,8 @@ cmi_ntv_hwdisable_mce(cmi_hdl_t hdl)
 		cmi_ntv_hwdisable_mce_xc();
 	} else {
 		CPUSET_ONLY(set, cp->cpu_id);
-		xc_call(NULL, NULL, NULL, CPUSET2BV(set),
-		    (xc_func_t)cmi_ntv_hwdisable_mce_xc);
+		xc_call((uintptr_t)NULL, (uintptr_t)NULL, (uintptr_t)NULL,
+		    CPUSET2BV(set), (xc_func_t)cmi_ntv_hwdisable_mce_xc);
 	}
 }
